src/hausmans/public_app/forms.py
from django import forms
from captcha.fields import ReCaptchaField


class ContactUsForm(forms.Form):
    first_name = forms.CharField()
    last_name = forms.CharField()
    email_address = forms.EmailField()
    message = forms.CharField(widget=forms.Textarea)
    captcha = ReCaptchaField(label='')
