requirements.txt
Django==3.1
gunicorn==20.0.4
whitenoise==5.0.1
django-recaptcha==2.0.6
gmail==0.6.3
